update documentation on network manager
